pkg-size: a package size calculator for OpenWrt
The package size calculator measures the uncompressed footprint
of a package and its dependencies.  It recurses down through each
dependency, so you can see all the files required by package. It
has options to ignore specific packages or files, and the ability
to ignore dependencies shared with other packages (to see the effect
of removing just the given one).

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>